more API

+++ server/src/controllers/relative.js
@@ -1,22 +1,23 @@
+const { validationResult } = require("express-validator");
 const Relative = require("../models/relative");
 const sendResponse = require("../utils/sendResponse");
 
 const getRelatives = async (req, res) => {
   try {
-    const relatives = await Relative.find({});
-    console.log(relatives);
-
-    if (relatives.length === 0) {
-      return sendResponse("No relatives found", res, 404);
-    }
+    const relatives = await Relative.find({ userId: req.user.id });
 
     res.status(200).json({ relatives });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
 const addRelative = async (req, res) => {
+  const errors = validationResult(req);
+  if (!errors.isEmpty()) {
+    return sendResponse(errors.array(), res, 400);
+  }
+
   try {
     req.body.userId = req.user.id;
 
@@ -26,42 +27,59 @@ const addRelative = async (req, res) => {
 
     res.json({ newRelative });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
 const updateRelative = async (req, res) => {
   try {
-    const relative = await Relative.findByIdAndUpdate(req.params.id, req.body);
+    const relative = await Relative.findOne({
+      _id: req.params.id,
+      userId: req.user.id,
+    });
 
     if (!relative) {
       return sendResponse("Unable to update relative", res, 404);
     }
 
-    res.json({ ...relative._doc, ...req.body });
+    const { firstName, lastName, email, mobileNumber } = req.body;
+    if (firstName) relative.firstName = firstName;
+    if (lastName) relative.lastName = lastName;
+    if (email) relative.email = email;
+    if (mobileNumber) relative.mobileNumber = mobileNumber;
+
+    await relative.save();
+
+    res.json({ relative });
   } catch (err) {
     return sendResponse(err.message, res);
   }
 };
 
-// delete single relative
 const deleteRelative = async (req, res) => {
   try {
-    const relative = await Relative.findByIdAndDelete(req.params.id);
+    const relative = await Relative.findOne({
+      _id: req.params.id,
+      userId: req.user.id,
+    });
 
     if (!relative) {
       return sendResponse("Unable to delete relative", res, 404);
     }
+
+    await relative.delete();
     sendResponse("Relative deleted", res, 200);
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
-// get single relative
 const getRelative = async (req, res) => {
   try {
-    const relative = await Relative.findById(req.params.id);
+    const relative = await Relative.findOne({
+      _id: req.params.id,
+      userId: req.user.id,
+    });
 
     if (!relative) {
       return sendResponse("No relative found", res, 404);
@@ -69,20 +87,20 @@ const getRelative = async (req, res) => {
 
     res.json({ relative });
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
 
 const deleteAllRelative = async (req, res) => {
   try {
-    const relative = await Relative.deleteMany({});
+    const relative = await Relative.deleteMany({ userId: req.user.id });
 
     if (!relative) {
       return sendResponse("Unable to delete relatives", res, 404);
     }
     sendResponse("All Relatives deleted", res, 200);
   } catch (err) {
-    return sendResponse("Server Error", res);
+    return sendResponse(err.message, res);
   }
 };
